c++ project
I have developed a basic DJ application called Otodecks. For the end of term assessment, you are tasked with developing the application further by adding a custom deck control Component and a music library Component, then integrating them into a new GUI layout of your own design. The custom deck control Component should have custom graphics, implemented in the paint function and it should offer a means to control a deck in some interesting way of your own design. The music library component should allow the user to manage a library of music within the application. They should be able to search the music library and load music from it into the decks. It should also persist between application loads, so it will need to store its state in a data file.